fixes to printf_functions.c
