+++ main.h
@@ -19,8 +19,8 @@ int _strlen(const char *s);
 int _printf(const char *format, ...);
 int print_string(char *s);
 int print_integer(long int num);
-double convert_negative_to_positive(double n);
-double reverse_number(double n);
+long int convert_negative_to_positive(long int n);
+long int reverse_number(long int n);
 char *_strcpy(char *dest, const char *src);
 char *_memset(char *s, char b, unsigned int n);
 int print_character(int idx, char *s);
